refactor(android): organize BLE, batch, and limitless sources into packages
Moves-only: com.friend.ios stays the root for generated Pigeon code and
standalone plugins; the BLE transport/services move to com.friend.ios.ble,
the batch capture sinks to com.friend.ios.batch, and the pendant codec +
drain engine to com.friend.ios.limitless. Manifest service names updated;
no logic changes.
