+++ app/android/app/src/main/kotlin/com/friend/ios/batch/BaseBatchAudioWriter.kt
@@ -1,4 +1,7 @@
-package com.friend.ios
+package com.friend.ios.batch
+
+import com.friend.ios.*
+import com.friend.ios.ble.OmiBleManager
 
 import android.content.Context
 import android.util.Log
