+++ app/android/app/src/main/kotlin/com/friend/ios/ble/BleHostApiImpl.kt
@@ -1,4 +1,6 @@
-package com.friend.ios
+package com.friend.ios.ble
+
+import com.friend.ios.*
 
 import android.app.Activity
 import android.bluetooth.BluetoothAdapter
