+++ app/android/app/src/main/kotlin/com/friend/ios/ble/OmiBleForegroundService.kt
@@ -1,4 +1,10 @@
-package com.friend.ios
+package com.friend.ios.ble
+
+import com.friend.ios.*
+import com.friend.ios.batch.LimitlessBatchAudioWriter
+import com.friend.ios.batch.OmiBackgroundAudioStreamer
+import com.friend.ios.batch.OmiBatchAudioWriter
+import com.friend.ios.limitless.LimitlessFlashDrainEngine
 
 import android.annotation.SuppressLint
 import android.app.*
