+++ app/ios/Runner/Ble/OmiBleManager.swift
